Requisito 4 FINALMENTE

+++ messages/disconnect.js
@@ -4,10 +4,12 @@ const users = require('../models/usersObject');
 module.exports = async (io) => {
   io.on('connection', async (socket) => {
     socket.on('disconnecting', () => {
-      const { nickname } = users[socket.id];
-      delete users[socket.id];
-      deleteOne(socket.id);
-      io.emit('logoff', nickname);
+      if (Object.values(users).length) {
+        const { nickname } = users[socket.id];
+        io.emit('logoff', nickname);
+        delete users[socket.id];
+        deleteOne(socket.id);
+      }
     });
   });
 };
